Add files via upload
This worksheet is called Summary and is in the workbook Quarter One Report.xlsx. In this worksheet, you’ll need to complete the following actions:

Create formulas that show the total quarter-one sales for both 2022 and 2023.

Create formulas that show the percentage increase in sales in 2023. 

And break down these totals by month with the use of further calculations.